requirements.txt
pyperclip
pyinstaller
pathspec
rich
python-magic